1.某些 SDK 底层含有 app_config.h，这样应用中自动生成的 app_config.h 就会遮盖它，导致编译报错，这里增加了可选的app_config_before.in 和 app_config_after.in 用于在生成的 app_config.h 中插入语句（我们只要插入 SDK 底层的 app_config.h 的 include 语句就行）

+++ components/my_file/my_file.py
@@ -35,7 +35,8 @@ def my_file_copy_one_kind_files_to(src_paths,kind,dst_path):
 # 将一个目录复制到另一个目录，如果另一个目录已经存在，则删除之
 def my_file_copy_dir_to(from_path,dst_path):
     my_file_rm_dir(dst_path) 
-    shutil.copytree(from_path,dst_path)
+    if os.path.exists(from_path):
+        shutil.copytree(from_path,dst_path)
 
 # 将一个目录下的所有内容复制到另一个文件夹，如果另一个目录不存在，则创建
 def my_file_copy_dir_contents_to(from_path,dst_path):
